separate address_standardization_processor.py
