+++ spark_pipeline_framework/transformers/elasticsearch_sender/v2/test/test_elasticsearch_sender.py
@@ -1,6 +1,7 @@
 from typing import List, Tuple
 
 import pytest
+from aioresponses import aioresponses
 from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.types import StructType, StructField, StringType
 
@@ -10,8 +11,111 @@ from spark_pipeline_framework.transformers.elasticsearch_sender.v2.elasticsearch
 )
 
 
+async def test_transform_async(spark_session: SparkSession) -> None:
+    with ProgressLogger() as progress_logger:
+        data = [
+            {"id": 1, "name": "Alice"},
+            {"id": 2, "name": "Bob"},
+        ]
+        schema = StructType(
+            [
+                StructField("id", StringType()),
+                StructField("name", StringType()),
+            ]
+        )
+        df = spark_session.createDataFrame(data, schema=schema)  # type: ignore[type-var]
+        df.createOrReplaceTempView("test_view")
+        sender = ElasticSearchSender(
+            index="test_index",
+            file_path=None,
+            view="test_view",
+            progress_logger=progress_logger,
+            run_synchronously=True,
+        )
+
+        with aioresponses() as m:
+            return_payload = {
+                "items": [
+                    {
+                        "index": {
+                            "_index": "test",
+                            "_id": "1",
+                            "_version": 1,
+                            "result": "created",
+                            "_shards": {"total": 2, "successful": 2, "failed": 0},
+                            "status": 201,
+                            "_seq_no": 0,
+                            "_primary_term": 1,
+                        }
+                    },
+                    {
+                        "index": {
+                            "_index": "test",
+                            "_id": "2",
+                            "_version": 1,
+                            "result": "created",
+                            "_shards": {"total": 2, "successful": 2, "failed": 0},
+                            "status": 201,
+                            "_seq_no": 0,
+                            "_primary_term": 1,
+                        }
+                    },
+                ],
+                "errors": False,
+            }
+
+            m.post(
+                "https://elasticsearch:9200/test_index/_bulk",
+                status=200,
+                payload=return_payload,
+            )
+
+            result_df = await sender._transform_async(df)
+            result_df.show(truncate=False)
+            assert result_df.count() == 1
+            assert result_df.collect()[0]["success"] == 2
+            assert result_df.collect()[0]["failed"] == 0
+
+
+async def test_transform_async_with_errors(spark_session: SparkSession) -> None:
+    with ProgressLogger() as progress_logger:
+        data = [
+            {"id": 1, "name": "Alice"},
+            {"id": 2, "name": "Bob"},
+        ]
+        schema = StructType(
+            [
+                StructField("id", StringType()),
+                StructField("name", StringType()),
+            ]
+        )
+        df = spark_session.createDataFrame(data, schema=schema)  # type: ignore[type-var]
+        df.createOrReplaceTempView("test_view")
+        sender = ElasticSearchSender(
+            index="test_index",
+            file_path=None,
+            view="test_view",
+            progress_logger=progress_logger,
+            run_synchronously=True,
+        )
+
+        with aioresponses() as m:
+            m.post(
+                "https://elasticsearch:9200/test_index/_bulk",
+                status=500,
+                payload={"error": "Internal Server Error"},
+            )
+
+            with pytest.raises(Exception):
+                result_df = await sender._transform_async(df)
+                result_df.show(truncate=False)
+                assert result_df.count() == 1
+                assert result_df.collect()[0]["success"] == 0
+                assert result_df.collect()[0]["failed"] == 2
+
+
 @pytest.mark.parametrize("run_synchronously", [True, False])
-def test_elasticsearch_sender(
+def test_elasticsearch_sender_real_elasticsearch(
     spark_session: SparkSession, run_synchronously: bool
 ) -> None:
     # Given
